+++ src/ScalarCollection/IntNullableCollection.php
@@ -43,12 +43,12 @@ class IntNullableCollection extends AbstractCollection implements ScalarCollecti
         return parent::toArray();
     }
 
-    protected function canAddValue(mixed $offset, mixed $value): bool
+    protected function canAddValue(mixed $value): bool
     {
         if (is_null($value) === false && is_int($value) === false) {
             throw new InvalidTypeException('$value should be of type int or null.');
         }
 
-        return parent::canAddValue($offset, $value);
+        return parent::canAddValue($value);
     }
 }